updated build scripts

+++ powerbi_powerbirs/BuildTurboImage.ps1
@@ -65,20 +65,19 @@ $Installer = "$DownloadPath\$InstallerName"
 if (!(Test-Path $Installer)) { 
 
     # Copy the sikulix resources folder to the desktop
-    Remove-Item -Path "$DesktopPath\Sikulix" -Recurse -Force
+    if (Test-Path "$DesktopPath\Sikulix") {
+        Remove-Item -Path "$DesktopPath\Sikulix" -Recurse -Force
+        }
     Copy-Item "$SupportFiles\Sikulix" -Destination $DesktopPath -Recurse -Force
-
-    # Wait for the warm up of the VM
-    Start-Sleep -Seconds 30
 
     # Pull down the sikulix and openjdk turbo images from turbo.net hub if they are not already part of the image
     $turboArgs = "config --domain=turbo.net"
     $ProcessExitCode = RunProcess "turbo.exe" $turboArgs $True
-    $turboArgs = "pull xvm,base,sikulix/sikulixide,microsoft/openjdk"
+    $turboArgs = "pull xvm,base,sikulix/sikulixide,microsoft/openjdk,isolate-edge-wc"
     $ProcessExitCode = RunProcess "turbo.exe" $turboArgs $True
 
     # Launch SikulixIDE to get the latest version
-    $turboArgs = "try sikulixide --using=microsoft/openjdk --offline --disable=spawnvm --isolate=merge-user --startup-file=java -- -jar @SYSDRIVE@\SikulixIDE\sikulixide-2.0.5.jar -r $sikulixPath\build.sikuli -f $env:userprofile\desktop\build-sikulix-log.txt"
+    $turboArgs = "try sikulixide,isolate-edge-wc --using=microsoft/openjdk --offline --disable=spawnvm --isolate=merge-user --startup-file=java -- -jar @SYSDRIVE@\SikulixIDE\sikulixide-2.0.5.jar -r $sikulixPath\build.sikuli -f $env:userprofile\desktop\build-sikulix-log.txt"
     $ProcessExitCode = RunProcess "turbo.exe" $turboArgs $True
     CheckForError "Checking process exit code:" 0 $ProcessExitCode $True # Fail on install error
 }
